Refactoring and fixing error regarding unknown data type
