+++ web/src/features/user/components/academic-experience.tsx
@@ -1,4 +1,4 @@
-const AcademicExperience = ({ data }: { data: unknown }) => {
+const AcademicExperience = ({ data }: { data: any }) => {
   return (
     <div className="bg-secondary px-4 py-3 sm:grid sm:grid-cols-2 sm:gap-4 sm:px-6 space-x-8">
       <dt className="text-lg font-bold font-['Montserrat'] text-gray-900">
